Configure kadi logger globally + other refactoring
