+++ starcheck/src/starcheck.pl
@@ -46,6 +46,8 @@ from starcheck.pcad_att_check import check_characteristics_date
 from starcheck.utils import (_make_pcad_attitude_check_report,
                              plot_cat_wrapper,
                              date2time, time2date,
+                             config_logging,
+                             set_kadi_scenario_default,
                              ccd_temp_wrapper,
                              starcheck_version, get_data_dir,
                              get_chandra_models_version,
@@ -71,6 +73,7 @@ my %par = (dir  => '.',
                    agasc_file => "${SKA}/data/agasc/proseco_agasc_1p7.h5",
 		   config_file => "characteristics.yaml",
 		   fid_char => "fid_CHARACTERISTICS",
+           verbose => 1,
 		   );
 
 
@@ -83,6 +86,7 @@ GetOptions( \%par,
 			'text!',
 			'yaml!',
 			'vehicle!',
+            'verbose=s',
 			'agasc_file=s',
 			'sc_data=s',
 			'fid_char=s',
@@ -106,6 +110,9 @@ my $font_stop = qq{</font>};
 
 usage( 1 )
     if $par{help};
+
+config_logging($STARCHECK, $par{verbose}, "kadi");
+set_kadi_scenario_default();
 
 # Find backstop, guide star summary, OR, and maneuver files.
 my %input_files = ();
