+++ starcheck/utils.py
@@ -1,27 +1,32 @@
-import os
 import functools
+import logging
+import os
+
 import numpy as np
-
-from Chandra.Time import DateTime
-from Chandra.Time import secs2date as time2date, date2secs as pydate2secs
-import starcheck
-from starcheck.pcad_att_check import make_pcad_attitude_check_report, check_characteristics_date
-from starcheck.calc_ccd_temps import get_ccd_temps
-from starcheck.plot import make_plots_for_obsid
-from starcheck.check_ir_zone import ir_zone_ok
-from starcheck import __version__ as version
-from kadi.commands import states
 import proseco.characteristics as proseco_char
+from Chandra.Time import DateTime
+from Chandra.Time import date2secs, secs2date
+from kadi.commands import states
+from testr import test_helper
+
+import starcheck
+from starcheck import __version__ as version
+from starcheck.calc_ccd_temps import get_ccd_temps
+from starcheck.check_ir_zone import ir_zone_ok
+from starcheck.pcad_att_check import make_pcad_attitude_check_report
+from starcheck.plot import make_plots_for_obsid
 
 
-def print_traceback_on_exception(func):
-    """Decorator to print a stack trace on exception.
+def python_from_perl(func):
+    """Decorator to facilitate calling Python from Perl inline.
 
-    The default perl inline suppresses this stack trace which makes debugging
-    difficult.
+    - Convert byte strings to unicode.
+    - Print stack trace on exceptions which perl inline suppresses.
     """
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
+        args = de_bytestr(args)
+        kwargs = de_bytestr(kwargs)
         try:
             return func(*args, **kwargs)
         except Exception:
@@ -46,55 +51,62 @@ def de_bytestr(data):
     return data
 
 
-@print_traceback_on_exception
-def date2time(date):
-    return pydate2secs(de_bytestr(date))
+date2time = python_from_perl(date2secs)
+time2date = python_from_perl(secs2date)
 
 
-@print_traceback_on_exception
+@python_from_perl
 def ccd_temp_wrapper(kwargs):
-    return get_ccd_temps(**de_bytestr(kwargs))
+    return get_ccd_temps(**kwargs)
 
 
-@print_traceback_on_exception
+@python_from_perl
 def plot_cat_wrapper(kwargs):
-    return make_plots_for_obsid(**de_bytestr(kwargs))
+    return make_plots_for_obsid(**kwargs)
 
 
-@print_traceback_on_exception
+@python_from_perl
 def starcheck_version():
     return version
 
 
-@print_traceback_on_exception
+@python_from_perl
 def get_chandra_models_version():
     return proseco_char.chandra_models_version
 
 
-@print_traceback_on_exception
+@python_from_perl
+def set_kadi_scenario_default():
+    # For kadi commands v2 running on HEAD set the default scenario to flight.
+    # This is aimed at running in production where the commands archive is
+    # updated hourly. In this case no network resources are used.
+    if test_helper.on_head_network():
+        os.environ.setdefault('KADI_SCENARIO', 'flight')
+
+
+@python_from_perl
 def get_kadi_scenario():
     return os.getenv('KADI_SCENARIO', default="None")
 
 
-@print_traceback_on_exception
+@python_from_perl
 def get_data_dir():
     sc_data = os.path.join(os.path.dirname(starcheck.__file__), 'data')
     return sc_data if os.path.exists(sc_data) else ""
 
 
-@print_traceback_on_exception
+@python_from_perl
 def _make_pcad_attitude_check_report(kwargs):
-    return make_pcad_attitude_check_report(**de_bytestr(kwargs))
+    return make_pcad_attitude_check_report(**kwargs)
 
 
-@print_traceback_on_exception
+@python_from_perl
 def make_ir_check_report(kwargs):
-    return ir_zone_ok(**de_bytestr(kwargs))
+    return ir_zone_ok(**kwargs)
 
 
-@print_traceback_on_exception
+@python_from_perl
 def get_dither_kadi_state(date):
-    date = date.decode('ascii')
     cols = ['dither', 'dither_ampl_pitch', 'dither_ampl_yaw',
             'dither_period_pitch', 'dither_period_yaw']
     state = states.get_continuity(date, cols)
@@ -107,7 +119,7 @@ def get_dither_kadi_state(date):
     return state
 
 
-@print_traceback_on_exception
+@python_from_perl
 def get_run_start_time(run_start_time, backstop_start):
     """
     Determine a reasonable reference run start time based on the supplied
@@ -128,9 +140,6 @@ def get_run_start_time(run_start_time, backstop_start):
     :returns: YYYY:DOY string of reference run start time
     """
 
-    run_start_time = de_bytestr(run_start_time)
-    backstop_start = de_bytestr(backstop_start)
-
     # For the special case where run_start_time casts as a float
     # check to see if it is negative and if so, set the reference
     # time to be a time run_start_time days back from backstop start
@@ -145,3 +154,41 @@ def get_run_start_time(run_start_time, backstop_start):
         else:
             raise ValueError("Float run_start_time should be negative")
     return ref_time.date
+
+
+@python_from_perl
+def config_logging(outdir, verbose, name):
+    """Set up file and console logger.
+    See http://docs.python.org/library/logging.html
+              #logging-to-multiple-destinations
+    """
+    # Disable auto-configuration of root logger by adding a null handler.
+    # This prevents other modules (e.g. Chandra.cmd_states) from generating
+    # a streamhandler by just calling logging.info(..).
+    class NullHandler(logging.Handler):
+        def emit(self, record):
+            pass
+    rootlogger = logging.getLogger()
+    rootlogger.addHandler(NullHandler())
+
+    loglevel = {0: logging.CRITICAL,
+                1: logging.INFO,
+                2: logging.DEBUG}.get(int(verbose), logging.INFO)
+
+    logger = logging.getLogger(name)
+    logger.setLevel(loglevel)
+
+    # Remove existing handlers if calc_ccd_temps is called multiple times
+    for handler in list(logger.handlers):
+        logger.removeHandler(handler)
+
+    formatter = logging.Formatter('%(message)s')
+
+    console = logging.StreamHandler()
+    console.setFormatter(formatter)
+    logger.addHandler(console)
+
+    filehandler = logging.FileHandler(
+        filename=os.path.join(outdir, 'run.dat'), mode='w')
+    filehandler.setFormatter(formatter)
+    logger.addHandler(filehandler)
